README.md
GRACE KLEINHELTER: Kleinhelter-Gallery-Online

Welcome to my project for Code Louisville’s Southern Indiana Front-End Web Development, course 1. 

Description:

My project is a a website for my family’s fine art gallery: Kleinhelter Gallery. As such, it features mostly photos, clickable and scrolling galleries to preview artist’s work, as well as a list of our regular artists, and a contact form for visitors to join our mailing list to keep up to date with our current and upcoming shows.

Features:

In my site I have included the following from the list of CSS and Javascript Features:

A navigation menu that expands and collapses at desktop and mobile sizes, becoming a hamburger icon that drops down vertically upon clicking.
A little CSS animation on the hamburger icon flipping into an X shape when clicked to be a close out button.
Flexbox and CSS Grid in the styling of my image galleries, in particular the “Artists” page is a responsive CSS Grid layout with a little animation as the artists’ names drop down when the cursor hovers over the images. The display changes to a different number of columns as the screen shifts between mobile and desktop sizes. The “footer” also adjusted to fit the screen size, stacking at mobile size.
A functional contact form (using formspree to connect our gallery’s email to the site) that uses Javascript to indicate an email has either been successfully submitted or not. If you click the submit button without filling out the form, a red notice will appear saying an empty form cannot be submitted. If the email is successful a green “thank you!” Button appears and disappears on a timer. 
There are two types of image galleries on the landing page. The first opens upon clicking the “sneak peak” button. This is a CSS Grid gallery that expands when the cursor hovers over an image. When clicked, the full sized image opens using javascript. Clicking outside the image closes the full sized image. 
The second gallery opens upon clicking the “view show” buttons on the Past Shows show cards. These open horizontal sliding galleries with clickable arrows that go right and left, allowing visitors to view the images in a loop. 

Current Issues:

The Past Shows galleries are cut-off at mobile sizes and the full-sized images in the Sneak Peak gallery are also cut off at mobile size.  

All of the images on the site are my own, so some are too large and a little laggy.